refactor(entities/coupon): Coupon interface entity layer로 이동
