+++ src/advanced/__tests__/advanced.test.tsx
@@ -3,7 +3,7 @@ import { describe, expect, test } from "vitest";
 import { act, fireEvent, render, screen, within } from "@testing-library/react";
 import CartPage from "@/refactoring/pages/cart";
 import AdminPage from "@/refactoring/pages/admin";
-import { Coupon } from "../../types";
+import { Coupon } from "@/refactoring/entities/coupon/model";
 import {
   Product,
   createProduct,
