+++ src/origin/components/CartPage.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
-import { CartItem, Coupon } from "../../types.ts";
+import { CartItem } from "../../types.ts";
+import { Coupon } from "@/refactoring/entities/coupon/model.ts";
 import { Product } from "../../refactoring/entities/product/model/product.ts";
 
 interface Props {
