+++ src/refactoring/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import CartPage from "./pages/cart/index.tsx";
 import AdminPage from "./pages/admin/index.tsx";
-import { Coupon } from "../types.ts";
+import { Coupon } from "./entities/coupon/model.ts";
 import { Product } from "./entities/product/model/product.ts";
 import { useCoupons, useProducts } from "./hooks";
