+++ src/refactoring/entities/coupon/model.ts
@@ -1,0 +1,8 @@
+import type { DiscountType as CouponDiscountType } from "./constants";
+
+export interface Coupon {
+  name: string;
+  code: string;
+  discountType: CouponDiscountType;
+  discountValue: number;
+}
