+++ src/refactoring/pages/admin/index.tsx
@@ -1,6 +1,6 @@
 import { Product } from "@/refactoring/entities/product";
 import { Discount } from "@/refactoring/entities/product/model/discount";
-import { Coupon } from "@/types";
+import { Coupon } from "@/refactoring/entities/coupon/model";
 import { useState } from "react";
 
 interface Props {
